LIBITD-408. Initial implementation of Bootstrap pagination
Modified the index_action_buttons partial to show the pagination bar
if a "pagination_object" is supplied.

Renamed the "object" parameter for the index_action_buttons partial to
"model_type", and updated the views that use it.

Updated the integration tests to reflect the new pagination bar (ul
instead of div).

https://issues.umd.edu/browse/LIBITD-408