LIBITD-438. Added "run_as_user" block for testing
Created block method to enable tests to run code inside block with
"current_user" being set to the given user.

Ensures that current_user is reset to the default, even if an exception
occurs.

Replaced existing uses of "CASClient::Frameworks::Rails::Filter.fake"
with "run_as_user" block.

Removed redundant setting of user to and "admin" user, because that
is already the default.

https://issues.umd.edu/browse/LIBITD-438