LIBITD-438. Suppressed User "List All" button for non-admins
Both the "show" and "edit" pages for the User object contained a
"List All" button. Since non-admin users can't see the list of other
users, this button needed to be suppressed for non-admins.

The "List All" button is part of _form_action_buttons and
_show_action_buttons partials, changes for the User functionality
could affect the other pages. Fortunately, Pundit allows the policy,
in this case "index?" to be queried inside the partial to determine
if the button should be shown.

Added an "index?" method to the personnel_request_policy to return
true so that the "List All" button would continue to be displayed
on the personnel request pages (since even non-admins can list
personnel requests).

https://issues.umd.edu/browse/LIBITD-438