LIBITD-494. Reduce the exposure to N+1 and leverage AR/Ariel more
N+1 issues cause lots of queries to the db which impact performance. Also makes
it difficult to manipulate queies with the active record query builder.
This looks to tamp down on the queries by eager loading and allow for a cleaner
interface to the orm. Focuses a lot on the user, roles, and cut-off interactions.

https://issues.umd.edu/browse/LIBITD-494