LIBITD-441. Refactored department/unit Labor Request tests
Added "verify_options" method to personnel_requests_test_helper.rb,
and refactored labor request tests to use it.

Added additional tests to labor_request_controller_test, so that
failure paths for "new" and "update" would be triggered. This
ensures that the selectable_departments and selectable_units are
established for these routes as well, as the page cannot be built
without these variables.

https://issues.umd.edu/browse/LIBITD-441