LIBITD-520. Modified enums to use defined numbers
See https://bendyworks.com/rails-enum-sharp-knife/ -- summary is
that if number isn't provided, the _order_ of the fields indicates
the number. This means that, without the number, inserting a value
into the middle of the list will essentially change the meaning of
the existing numbers.

Providing the number ensures that the list remains stable, even as
new values (with their own numbers) are added.

https://issues.umd.edu/browse/LIBITD-520