LIBITD-525. Gave each model a "description" to use with confirm prompt
For models having a "name" or a "position_description", used an
"alias_attribute".

For other models having more complex needs, like Role, added a method
to generate the description text.

Considered centralizing all the descriptions into a single method in
application_helper.rb, but ultimately went with the alias_attribute or
description method to make it a little more obvious, especially if new
models are added (as opposed to making it auto-magical).

https://issues.umd.edu/browse/LIBITD-525