LIBITD-528. Set fields to default to "1" in "new" page.
Modified the ContractorRequest and LaborRequest models so that the
following fields default to "1":

ContractorRequests:
* number_of_months

LaborRequests:
* number_of_positions
* number_of_weeks

The changes in the "after_initialize" needed to be protected by
a "has_attribute" guard clause because LaborRequest is used in
"select" statements that don't include all the columns, and
some of the unit tests were failing (with
ActiveModel::MissingAttributeError). Put the same change into
ContractorRequest change for consistency.

See http://stackoverflow.com/questions/21792996/in-a-rails-activerecord-model-is-using-after-initialize-callbacks-a-very-bad-id
and http://stackoverflow.com/questions/6820156/activemodelmissingattributeerror-occurs-after-deploying-and-then-goes-away-aft

https://issues.umd.edu/browse/LIBITD-528