LIBITD-554. Missing department is validation error.
Previously, attempting to create a personnel request without a department was generating a Pundit::NotAuthorized exception that was then getting swallowed and the user would be sent back to the form with no explanatory text as to why their request was not created. I changed the personnel request policy to only check for department permissions if a department is set in the request. Otherwise, it lets it through to be caught by the validation routines. I also changed the "not_authorized" exception handlers in the personnel request controllers to properly redirect and display flash messages. Finally, I updated the integration tests for the permissions to access the labor request form, and added similar tests for the contractor and staff request forms.

https://issues.umd.edu/browse/LIBITD-554