LIBITD-565 Fix the responsive functionality on tables
This wraps the index table in a .table-responsive. It also adds a
common _new_or_edit partial in the shared directory that new or edit views can
use ( since both just render the form ).

This also adds a small feature to the umd_style gem. To get this working now, i've
included a local override of the template but am submitting a PR to the umd_style
repo.

https://issues.umd.edu/browse/LIBITD-565