LIBITD-536 More N+1 fixes and adding counter caches
There are several places that have the N+1 query problems as well as missing
counter cache columns. This attempts to fix this by optimizing the :includes, adding
counter cache, and removing unused eager loads.

Also includes the bullet gem, which will warn in dev and test mode when these issues
arrise in the code.

https://issues.umd.edu/browse/LIBITD-536