LIBITD-537. Updated report infrastructure to allow for error messages
Added "parameters_valid?" and "error_message" messages to the Reportable
interface, to allow for parameter error-checking.

Updated report_job.rb to use the "parameters_valid?" method of the
report, and set the error status and error message for invalid
parameters.

Updated the "show" template to display the error messsage, if present.

Updated the Reports.md documentation.

https://issues.umd.edu/browse/LIBITD-537