LIBITD-749 Make more obvious what user and org permissions are
This tries to make it more clean in UI what the current roles are in
relation to orgs and users. User detail pages show what orgs are
included in the user's roles ( all orgs via the hierachy ).

The same is done for orgs, with org detail pages showing what users have
access to the org going up the ancestor chain.

The goal is to make it more clear what users have access to what orgs.