LIBITD-934 make requests belong_to a dept and unit
This changes the generic organization to be a department, while allowing
a request to also belong_to a unit organization. The users can select a
department in the organization select box, and then can select a unit
(optional). The values visible in the unit select should have the parent
department that is selected in the organization select box.