LIBITD-970. Updated test fixtures
Need to update the test fixtures so that the "department" and "unit"
data was consistent for requests and archived_requests.

Recreated sample data and regenerates "archive_requests",
"organizations", and "requests.yml" test fixtures by running the
following commands:

```
> rake db:reset_with_sample_data

> rake db:migrate

> rake db:archive_current_records[2016]

> rake db:populate_sample_data

> rake db:migrate

> rm test/fixtures/archived_requests.yml test/fixtures/requests.yml test/fixtures/organizations.yml

> rake db:fixtures:dump

> rm test/fixtures/links.yml
```

Note: Needed to delete "links.yml" because the current test fixture
for the "links" table is named "link.yml". Added a Jira issue LIBITD-978
to fix this.

https://issues.umd.edu/browse/LIBITD-970