LIBITD-999. Fixed broken tests on feature/LIBITD-749 branch.
Pinned axlsx version to 2.0.1, as that more closely matches what was in
v1.2.0 (which used v2.0.0).

Unpinned rubyzip version, as axlsx needs it to be "~> 1.0.0".

Pinned "roo" version to 1.13.2, as that matches what was in v1.2.0.

Added subdirectories of "app/models" directory to the autoload path. This
was needed for the "report" tests to find the "report" models.

https://issues.umd.edu/browse/LIBITD-999