LIBITD-1083 Make Unit required for Unit level users
This requires the unit field to be required for users with only unit
level permissions. The request policy now checks if the user has the
correct permissions to create the record. If the user is attempting to
create a record that's not tied to their unit, an error is displayed.

https://issues.umd.edu/browse/LIBITD-1083