LIBITD-1275. Fix Excel formatting for personnel request currency fields
Fixes an issue where the currency fields in the Excel spreadsheet for
exported personnel requests were displaying as a general text field,
instead of a more specific numeric field.

Modified index.xlsx.axlsx to use workbook-defined styles, as in
other Axlsx views.

Modified XlsxHelper to use CURRENCY_FIELDS and LENGTHY_FIELDS constants
directly (as is done in other places in the code), instead of delegating
to methods in the RequestHelper module. Fixed the "field_formats" method
to actually return a format, instead of width. Also flipped the order of
the return values in the "formats_and_widths" method because that was
what seemed to be expected in other parts of the code (and the order
matches the method name).

Removed unused methods in the RequestHelper module
(app/helpers/request_helper.rb), including the generated "xlsx"-specific
methods for currency fields, which are no longer needed.

In labor_requests_controller_test.rb, removed the code that checked for
the "$" because the cell no longer contains a string (it contains a
Float instead). Looked through the Roo methods, but could not find a
way to check the style of the cells.

https://issues.umd.edu/browse/LIBITD-1275