LIBITD-1289. Updated tests to run in headless Chrome
In test/application_system_test_case.rb, modified the Capybara
configuration to enable files to be downloaded in a headless Chrome
browser. Use a modified version of the code found at

* https://gist.github.com/bbonamin/4b01be9ed5dd1bdaf909462ff4fdca95

which is similar to the answer suggested in

* https://stackoverflow.com/a/47649592

Also removed the "require" statements for"action_pack" and "fileutils",
as they did not seem to be necessary.

In test/system/index_view_test.rb, modified the "mkdir" call to use the
"FileUtils" class, because the method is not available on the "File"
class (and it seemed like that was what was intended, given the
"require" statement for "fileutils").

https://issues.umd.edu/browse/LIBITD-1289