LIBITD-1304. Fixed "Reset Sorting" link on "My Requests" page
Fixed an issue in which the "Reset Sorting" link on the "My Requests"
was resetting to the application home page, instead of the "My Request"
page.

Modified the "reset_sorts_link" method in the
app/helpers/request_helper.rb to set the "Reset Sorting Link" to the
"my_requests_path", as the model class appears to be "Request" when the
"My Requests" page is being displayed.

https://issues.umd.edu/browse/LIBITD-1304