LIBITD-1659. Fixed "money" gem deprecation warnings
Fixed deprecation warnings generated by the "money" gem regarding
changes to the defaults in the next version. Fixed the warnings
by setting the defaults in the next version in the configuration file.

See https://github.com/RubyMoney/money#deprecation

These changes are not expected to change the behavior, because we do
not use the "money" gem that extensively.

https://issues.umd.edu/browse/LIBITD-1659